debian-4.9.0: Tweak untangle patches to fix compiler warnings
Fix indentation in udp.c to fix misleading indentation warning.

Use the already deconstified pskb pointer in xt_socket to restore
the mark.  This fixes a discarded qualifier warning.

Initialize outdev to NULL to fix a possible use of uninitialized variable
warning.

The 4.9 kernel now builds with only 1 warning (which was introduced
by debian patches and is benign).

NGFW-11556 #code-completed