linksys/omnia: Have the kernel mount the rootfs as rw
Currently we are depending on 01minwrites to remount the rootfs as rw.  But
systemd expects the filesystem to be rw before that.  We could let systemd
remount the rootfs by supplying an fstab file for each device, or we can just
tell the kernel to mount the rootfs as rw from the beginning.