debian-4.19.0: Tweak iptables socket match hack
Only or in the upper two bytes of the socket mark.  If the interface
marks in the packet don't match the interface marks in the socket
mark, we may cause weird behaviour.  Since the purpose of this hack
is to restore the qos/bandwidth control mark, don't even mess with
the interface bytes.

NGFW-12813