debian-4.9.0: Remove socket mark orring with packet mark from extensions patch
The mark field of the sk_buff structure is a union with reserved_tailroom.
Since reserved_tailroom may be used while creating tcp packets, the skb->mark
field may be a non-zero value when it reaches __ip_queue_xmit and
__ip_make_skb.  By orring the socket mark instead of setting it, we are
allowing the reserved_tailroom value to propogate further down the networking
stack as the packet mark, which is obviously invalid.  Restore the kernel's
original intent to reinitialize the packet mark in these functions by setting
the socket mark instead of orring.

The original intent of the change to orring in the socket mark is not known.
We are guessing that at the time it was written (which was 2.6.26 timeframe
or earlier) the kernel, or perhaps a different Untangle kernel patch, was
manipulating the packet mark before it reached these functions, and that
those manipulations needed to be preserved even when applying the socket
mark.