docker: kernel builds occur in AWS (not in travis), where package-server can be directly resolved...
... but only on the host itself: from within docker, it gets
expanded/queried as package-server.untangle.com, which returns
the incorrect 173.14.28.193.

We move the fix to Jenkinsfile instead, so that:

  - any developer can build in docker, using the default public IP for
    package-server without passing an extra PACKAGE_SERVER_IP
    environment variable

  - jenkins-based builds do the right thing by using the internal IP

This is still sub-optimal, as it will require a change if/when we
relocate or reinstantiate package-server, but it can't really be helped:
the decision to use public EIPs for build machines doesn't leave room
for a cleaner solution at the moment.